created timer class, broke Run into smaller methods, let game continue on incorrect

+++ MathGame.kwm0304/Models/LeaderboardEntry.cs
@@ -2,14 +2,14 @@ namespace MathGame.CSA.Models;
 
 public class LeaderboardEntry
 {
-  public int EntryScore { get; set; }
+  public double EntryScore { get; set; }
   public string Initials { get; set; }
   public DateTime EntryDate { get; set; }
   public string EntryDifficulty { get; set; }
   public TimeSpan TimeTaken { get; set; }
   public bool Completed { get; set; }
 
-  public LeaderboardEntry(int entryScore, string initials, string entryDifficulty, TimeSpan timeTaken, bool completed)
+  public LeaderboardEntry(double entryScore, string initials, string entryDifficulty, TimeSpan timeTaken, bool completed)
   {
     EntryScore = entryScore;
     Initials = initials;
